renamed package into life_system
