+++ src/model/game_object/entity/SimpleEnemy.java
@@ -1,7 +1,8 @@
 package model.game_object.entity;
 
 import java.awt.Image;
-import model.life.LifeSystem;
+
+import model.life_system.LifeSystem;
 import model.movement.Movement;
 import model.weapon.Weapon;
 import utilities.Pair;
