+++ src/model/life_system/ExtendibleMaxHealthSystem.java
@@ -1,4 +1,4 @@
-package model.life;
+package model.life_system;
 
 /**
  *
